Add GraphQL Codegen documentation
Created a new file, codegen.md, which provides a detailed description and guidance on how to generate TypeScript type definitions and execution code from GraphQL schemas using GraphQL Code Generator. The newly added documentation includes installation and initialization instructions, as well as examples of usage. This was done to facilitate the understanding and utilization of GraphQL Code Generator for the team.